+++ config.go
@@ -31,15 +31,15 @@ const (
 	//degreeMax uint8 = 128
 	// 最大存储数，超过次数一律做新值换算
 	//lilyMax      uint32 = 4294967280
-	cityDistance uint32 = 268435455
+	cityDistance int64 = 268435455
 	// mallDistance level1间隔 ld1=(treeCount+1)/128=2097152 128^3
-	mallDistance uint32 = 2097152
+	mallDistance int64 = 2097152
 	// trolleyDistance level2间隔 ld2=(16513*127+1)/128=16384 128^2
-	trolleyDistance uint32 = 16384
+	trolleyDistance int64 = 16384
 	// purseDistance level3间隔 ld3=(129*127+1)/128=128 128^1
-	purseDistance uint32 = 128
+	purseDistance int64 = 128
 	// boxDistance level4间隔 ld3=(1*127+1)/128=1 128^0
-	boxDistance uint32 = 1
+	boxDistance int64 = 1
 )
 
 //const (
