+++ selector.go
@@ -280,7 +280,7 @@ func (s *Selector) shellDesc(is []interface{}) []interface{} {
 	return is
 }
 
-func (s *Selector) hashKeyFromValue(params []string, value interface{}) uint32 {
+func (s *Selector) hashKeyFromValue(params []string, value interface{}) int64 {
 	hashKey, support := s.getInterValue(params, value)
 	if !support {
 		return 0
@@ -289,7 +289,7 @@ func (s *Selector) hashKeyFromValue(params []string, value interface{}) uint32 {
 }
 
 // getInterValue 根据索引描述和当前检索到的value对象获取当前value对象所在索引的hashKey
-func (s *Selector) getInterValue(params []string, value interface{}) (hashKey uint32, support bool) {
+func (s *Selector) getInterValue(params []string, value interface{}) (hashKey int64, support bool) {
 	reflectObj := reflect.ValueOf(value) // 反射对象，通过reflectObj获取存储在里面的值，还可以去改变值
 	if reflectObj.Kind() == reflect.Map {
 		interMap := value.(map[string]interface{})
